+++ EvolutionaryGame/EvolutionaryGame/WeaponStats.h
@@ -12,11 +12,11 @@
 
 @property Boolean evolved;
 
-@property int fireDamage;
-@property int iceDamage;
-@property int lightningDamage;
-@property int earthDamage;
-@property int waterDamage;
+@property float fireDamage;
+@property float iceDamage;
+@property float lightningDamage;
+@property float earthDamage;
+@property float waterDamage;
 
 @property int killCount;
